# docs/swagger-doc.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Groccery API documentation",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://groccery-api.onrender.com/api",
      "description": "Production server"
    },
    {
      "url": "http://localhost:4000/api",
      "description": "Local development server"
    }
  ],
  "paths": {
...
    "/address": {
      "get": {
        "security": [{ "xAuth": [] }],
        "tags": ["Address"],
        "summary": "Get a list of addresses",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": { "type": "string", "default": "success" },
                    "count": { "type": "number" },
                    "data": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/Address" }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Authorization Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Token error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "post": {
        "security": [{ "xAuth": [] }],
        "tags": ["Address"],
        "summary": "Create a new address",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "fullName": { "type": "string", "default": "Jon Doe" },
                  "phone": { "type": "string", "default": "5265895475" },
                  "pincode": { "type": "string", "default": "526854" },
                  "city": { "type": "string", "default": "New york" },
                  "house": { "type": "string", "default": "Bakers street 315" },
                  "area": { "type": "string", "default": "wallstreet" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": { "type": "string", "default": "success" },
                    "message": {
                      "type": "string",
                      "default": "Address added successfully."
                    },
                    "data": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/Address" }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Authorization Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Token error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },

          "422": {
            "description": "Validaton error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/address/{id}": {
      "get": {
        "security": [{ "xAuth": [] }],
        "tags": ["Address"],
        "summary": "Get a address with specified ID",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "Address ID",
            "type": "string",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": { "type": "string", "default": "success" },
                    "data": {
                      "type": "object",
                      "$ref": "#/components/schemas/Address"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Authorization Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Token error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "delete": {
        "security": [{ "xAuth": [] }],
        "tags": ["Address"],
        "summary": "Delete a address with specified ID",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "Address ID",
            "type": "string",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": { "type": "string", "default": "success" },
                    "message": {
                      "type": "string",
                      "default": "Address deleted successfully."
                    },
                    "data": {
                      "type": "object",
                      "$ref": "#/components/schemas/Address"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Authorization Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Token error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "put": {
        "security": [{ "xAuth": [] }],
        "tags": ["Address"],
        "summary": "Update a Address",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "Address ID",
            "type": "string",
            "required": true
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "fullName": { "type": "string", "default": "Jon Doe" },
                  "phone": { "type": "string", "default": "6854585256" },
                  "pincode": { "type": "string", "default": "253641" },
                  "city": { "type": "string", "default": "New york" },
                  "house": { "type": "string", "default": "Bakers street 315" },
                  "area": { "type": "string", "default": "wallstreet" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": { "type": "string", "default": "success" },
                    "message": {
                      "type": "string",
                      "default": "Address updated successfully."
                    },
                    "data": {
                      "type": "object",
                      "$ref": "#/components/schemas/Address"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Authorization Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Token error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "xAuth": {
        "type": "apiKey",
        "name": "x-token",
        "in": "header"
      }
    },
    "schemas": {
...
      "Address": {
        "properties": {
          "userId": { "type": "string" },
          "fullName": { "type": "string" },
          "phone": { "type": "string" },
          "pincode": { "type": "string" },
          "city": { "type": "string" },
          "house": { "type": "string" },
          "area": { "type": "string" },
          "createdAt": { "type": "string", "format": "date-time" },
          "modifiedAt": { "type": "string", "format": "date-time" },
          "_id": { "type": "string" }
        }
      },
...
      "Error": {
        "properties": {
          "status": { "type": "string", "default": "error" },
          "message": { "type": "string" }
        }
      }
    }
  }
}